// src/store/role/Types.tsx
export enum Resource {
  PROJECT = 'PROJECT',
  MONITORING = 'MONITORING',
  FLOW = 'FLOW',
  KAFKA = 'KAFKA',
  GROUP = 'GROUP',
  LOGGING = 'LOGGING',
  TASKS = 'TASKS',
  TRACING = 'TRACING',
  FULL_TEXT_INDEX = 'FULL_TEXT_INDEX',
  PIPELINE = 'PIPELINE',
  DICTIONARY = 'DICTIONARY',
  ARCHIVE = 'ARCHIVE',
}

export enum Role {
  KAFKA_VIEWER = 'KAFKA_VIEWER',
  KAFKA_EDITOR = 'KAFKA_EDITOR',
  MON_VIEWER = 'MON_VIEWER',
  MON_EDITOR = 'MON_EDITOR',
  FLOW_VIEWER = 'FLOW_VIEWER',
  FLOW_EDITOR = 'FLOW_EDITOR',
  TRACING_EDITOR = 'TRACING_EDITOR',
  GROUP_MANAGER = 'GROUP_MANAGER',
  ACCESS_MANAGER = 'ACCESS_MANAGER',
  FULL_TEXT_INDEX_VIEWER = 'FULL_TEXT_INDEX_VIEWER',
  FULL_TEXT_INDEX_EDITOR = 'FULL_TEXT_INDEX_EDITOR',
  FULL_TEXT_INDEX_EXPORT = 'FULL_TEXT_INDEX_EXPORT',
  DICTIONARY_VIEWER = 'DICTIONARY_VIEWER',
  DICTIONARY_EDITOR = 'DICTIONARY_EDITOR',
  LOOKUP_MANAGER = 'LOOKUP_MANAGER',
  ARCHIVE_VIEWER = 'ARCHIVE_VIEWER',
  ARCHIVE_EDITOR = 'ARCHIVE_EDITOR',
  ARCHIVE_INDEX_EXPORT = 'ARCHIVE_INDEX_EXPORT',
  AUDIT_EVENTS_EXPORT = 'AUDIT_EVENTS_EXPORT',
  SEARCH_PATTERN_VIEWER = 'SEARCH_PATTERN_VIEWER',
  SEARCH_PATTERN_EDITOR = 'SEARCH_PATTERN_EDITOR',
  DATA_VIEW_EDITOR = 'DATA_VIEW_EDITOR',
  INDEX_DATA_VIEW_PERMISSION_MANAGER = 'INDEX_DATA_VIEW_PERMISSION_MANAGER',
}

export enum PVMRole {
  ALMGR_RULE_VIEW = 'ALMGR_RULE_VIEW',
  ALMGR_RULE_EDIT = 'ALMGR_RULE_EDIT',
  ALMGR_RULE_TEST = 'ALMGR_RULE_TEST',
  ALMGR_GROUP_RULE_VIEW = 'ALMGR_GROUP_RULE_VIEW',
  ALMGR_GROUP_RULE_EDIT = 'ALMGR_GROUP_RULE_EDIT',
  ALMGR_EVENTS_VIEW = 'ALMGR_EVENTS_VIEW',
  NTFMGR_CHANNEL_VIEW = 'NTFMGR_CHANNEL_VIEW',
  NTFMGR_CHANNEL_EDIT = 'NTFMGR_CHANNEL_EDIT',
  NTFMGR_TEMPLATE_VIEW = 'NTFMGR_TEMPLATE_VIEW',
  NTFMGR_TEMPLATE_EDIT = 'NTFMGR_TEMPLATE_EDIT',
  NTFMGR_GROUP_VIEW = 'NTFMGR_GROUP_VIEW',
  NTFMGR_GROUP_EDIT = 'NTFMGR_GROUP_EDIT',
  MONA_CONNECTIONS_EDITOR = 'MONA_CONNECTIONS_EDITOR',
  MONA_CONNECTIONS_VIEWER = 'MONA_CONNECTIONS_VIEWER',
  MONA_FILTER_VIEWER = 'MONA_FILTER_VIEWER',
  MONA_FILTER_EDITOR = 'MONA_FILTER_EDITOR',
  MONA_METADATA_VIEWER = 'MONA_METADATA_VIEWER',
  MONA_METADATA_EDITOR = 'MONA_METADATA_EDITOR',
  MONA_DATA_VIEWER = 'MONA_DATA_VIEWER',
  MONA_QUOTA_VIEWER = 'MONA_QUOTA_VIEWER',
  MONA_QUOTA_EDITOR = 'MONA_QUOTA_EDITOR',
  MONA_QUOTA_METRICS_WRITE = 'MONA_QUOTA_METRICS_WRITE',
  MONA_METRICS_EXTERNAL_PUSHER = 'MONA_METRICS_EXTERNAL_PUSHER',
  OSIRIS_OBJECT_VIEW = 'OSIRIS_OBJECT_VIEW',
  OSIRIS_OBJECT_EDIT = 'OSIRIS_OBJECT_EDIT',
  OSIRIS_OBJECT_CREATE = 'OSIRIS_OBJECT_CREATE',
  OSIRIS_OBJECT_DELETE = 'OSIRIS_OBJECT_DELETE',
  OSIRIS_SETTINGS_VIEW = 'OSIRIS_SETTINGS_VIEW',
  OSIRIS_SETTINGS_EDIT = 'OSIRIS_SETTINGS_EDIT',
  BAM_CONNECTION_VIEWER = 'BAM_CONNECTION_VIEWER',
  BAM_CONNECTION_EDITOR = 'BAM_CONNECTION_EDITOR',
  BAM_CONNECTION_LOG_VIEWER = 'BAM_CONNECTION_LOG_VIEWER',
  BAM_PROCESS_VIEWER = 'BAM_PROCESS_VIEWER',
  BAM_PROCESS_DATA_VIEWER = 'BAM_PROCESS_DATA_VIEWER',
  BAM_FILTER_VIEWER = 'BAM_FILTER_VIEWER',
  BAM_FILTER_EDITOR = 'BAM_FILTER_EDITOR',
  BAM_UNDEFINED_CONNECTION_EDITOR = 'BAM_UNDEFINED_CONNECTION_EDITOR',
  AIOM_ALGORITHM_VIEWER = 'AIOM_ALGORITHM_VIEWER',
  AIOM_ALGORITHM_EDITOR = 'AIOM_ALGORITHM_EDITOR',
  AIOM_ALGORITHM_ADMIN = 'AIOM_ALGORITHM_ADMIN',
  TC_PIPELINE_VIEW = 'TC_PIPELINE_VIEW',
  TC_PIPELINE_ADMIN = 'TC_PIPELINE_ADMIN',
  FEATURE_SETTINGS_VIEW = 'FEATURE_SETTINGS_VIEW',
  FEATURE_SETTINGS_EDIT = 'FEATURE_SETTINGS_EDIT',
  PVM_SECRETS_VIEWER = 'PVM_SECRETS_VIEWER',
  PVM_SECRETS_EDITOR = 'PVM_SECRETS_EDITOR',
}

export enum Unit {
  USER = 'USER',
  GROUP = 'GROUP',
}

export const UNIT_MAP = {
  GROUP: 'Группа',
  USER: 'Пользователь',
};

export interface UnitRole {
  unit: Unit;
  unitId: number;
  roles: Array<Role>;
}